Add `--sane` option
`--sane` specified that is there are not enough date files to
complete the operation, still exist with a true status (0). This
is a common situation as this script will be used when new radials
are coming in. If there are not enough radials, we do not want to
get an error, but rather quit quietly. That is unlike other errors
where we should signal the error for further investigation.