Generate current files with correct permissions
Generate ACORN hourly average current files with 0444 permissions.
This is important when dealing with the pipeline, as it will
inject the file back into an incoming directory and might overwrite
a file that had been there already with 0444 permissions and may
or may not be accessed by another user (usually the talend user
which harvests the file).